Verificação de cadastro _nome_ e _senha_ á existente (Teste ≠ teste: diferentes)

+++ Model/Model_User.java
@@ -3,14 +3,16 @@ package Model;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Scanner;
+import java.io.BufferedReader;
 import java.io.Console;
+import java.io.FileReader;
 
 public class Model_User {
   private String nome;
   private String senha;
 
   public Model_User(Scanner scanner) {
-    System.out.print("Digite o nome do usuário: ");
+    System.out.print("Digite o nome do usuario: ");
     this.nome = scanner.nextLine();
 
     Console console = System.console();
@@ -26,12 +28,36 @@ public class Model_User {
 
   public void salvarEmArquivo() {
     try {
+      // 1. Verifica se o usuário já existe
+      BufferedReader reader = new BufferedReader(new FileReader("DB/usuarios.data"));
+      String linha;
+      boolean existe = false;
+
+      /// Verifica se o nome já existe no arquivo
+      while ((linha = reader.readLine()) != null) {
+        String[] partes = linha.split("\\|"); // separa nome e senha
+        if (partes[0].equals(nome) && partes[1].equals(senha)) { // compara o nome e senha digitado
+          existe = true;
+          break; // já achou, pode parar
+        }
+      }
+
+      reader.close();
+
+      if (existe) {
+        System.out.println("Erro: Usuario '" + nome + "' ja existe!\n");
+        return; // não salva
+      }
+
+      // 2. Salva se não existir
       FileWriter writer = new FileWriter("DB/usuarios.data", true); // true = append
       writer.write(nome + "|" + senha + "\n");
       writer.close();
+
       System.out.println("Usuario salvo com sucesso!\n");
+
     } catch (IOException e) {
-      System.out.println("Erro ao salvar usuario: " + e.getMessage());
+      System.out.println("Erro ao acessar o arquivo: " + e.getMessage());
     }
   }
